+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added ApiResultException class.
 
 ## [2.0.0] - 2018-05-14
 - **Backward compatibility break**: Changed parameter order in constructor in ApiResult class.