+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## [2.1.0] - 2019-06-14
 - Added ApiResultException class.
 - Added getJsonEncodeOptions method to ApiController class.
 - Throws exception if result could not be JSON-encoded.
@@ -10,4 +10,5 @@
 ## 1.1.0 - 2018-04-29
 - First stable revision.
 
+[2.1.0]: https://github.com/themichaelhall/bluemvc-api/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/themichaelhall/bluemvc-api/compare/v1.1.0...v2.0.0